Quill changes

+++ ContactsApp1/src/main/scala/BackEnd/entity/Contact.scala
@@ -1,3 +1,3 @@
 package BackEnd.entity
 
-case class Contact(id:BigInt , firstName:String , lastName:String , email:String, phone:String)
+case class Contact(id:Int , firstName:String , lastName:String , email:String, phone:String)
